final push for today
I added edit, add, delete animal. you can not only do that if you are logged in and only to the shelter you are logged in to. I added a logout that will only show up if you are logged in. I added the login and signup to every page as well as the logout. I put the shelter information, phone, email, and address under the shelter name.